requirements.txt
streamlit
tensorflow
pillow
numpy
flet
